requirement.txt
alabaster==0.7.12
anyio==3.6.2
argh==0.26.2
asn1crypto==1.3.0
astroid==2.4.2
astropy==4.0.1.post1
atomicwrites==1.4.0
attrs==19.3.0
autopep8==1.5.3
Babel==2.8.0
backcall==0.2.0
backports.functools-lru-cache==1.6.1
backports.shutil-get-terminal-size==1.0.0
backports.tempfile==1.0
backports.weakref==1.0.post1
beautifulsoup4==4.9.1
bitarray==1.4.0
bittrex-api==0.0.21
bkcharts==0.2
bleach==3.1.5
bokeh==2.1.1
boto==2.49.0
brotlipy==0.7.0
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
charset-normalizer==2.1.1
click==7.1.2
cloudpickle==1.5.0
colorama==0.4.3
contextlib2==0.6.0.post1
cryptography==2.9.2
cycler==0.10.0
Cython==0.29.21
cytoolz==0.10.1
dask==2.20.0
decorator==4.4.2
defusedxml==0.6.0
diff-match-patch==20200713
distributed==2.20.0
dnspython==2.2.1
docutils==0.16
email-validator==1.1.2
entrypoints==0.3
et-xmlfile==1.0.1
fake-useragent==0.1.14
fastapi==0.62.0
fastcache==1.1.0
filelock==3.0.12
flake8==3.8.3
Flask==1.1.2
fsspec==0.7.4
h5py==2.10.0
HeapDict==1.0.1
html5lib==1.1
idna==2.10
imageio==2.9.0
imagesize==1.2.0
importlib-metadata==1.7.0
intervaltree==3.0.2
ipykernel==5.3.2
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.17.1
jeepney==0.4.3
Jinja2==2.11.2
joblib==0.16.0
json5==0.9.5
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.6
jupyter-console==6.1.0
jupyter-core==4.6.3
jupyterlab==2.1.5
jupyterlab-server==1.2.0
kcu==0.0.73
keyring==21.2.1
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
lxml==4.5.2
MarkupSafe==1.1.1
matplotlib==3.2.2
mccabe==0.6.1
mistune==0.8.4
more-itertools==8.4.0
mpmath==1.1.0
msgpack==1.0.0
nbconvert==5.6.1
nbformat==5.0.7
networkx==2.4
nltk==3.5
noraise==0.0.16
nose==1.3.7
notebook==6.0.3
numba==0.50.1
olefile==0.46
openpyxl==3.0.4
packaging==20.4
parso==0.7.0
partd==1.1.0
path==13.1.0
pathlib2==2.3.5
pathtools==0.1.2
patsy==0.5.1
pep8==1.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
pip==20.1.1
pkginfo==1.5.0.1
pluggy==0.13.1
ply==3.11
prometheus-client==0.8.0
prompt-toolkit==3.0.5
psutil==5.7.0
ptyprocess==0.6.0
py==1.9.0
pycodestyle==2.6.0
pycosat==0.6.3
pycparser==2.20
pycurl==7.43.0.5
pydantic==1.10.2
pydocstyle==5.0.2
pyflakes==2.2.0
Pygments==2.6.1
PyJWT==1.7.1
pylint==2.5.3
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyrsistent==0.16.0
PySocks==1.7.1
pytest==5.4.3
python-dateutil==2.8.1
python-decouple==3.3
python-jsonrpc-server==0.3.4
python-language-server==0.34.1
pytz==2020.1
PyWavelets==1.1.1
pyxdg==0.26
PyYAML==5.3.1
pyzmq==19.0.1
QDarkStyle==2.8.1
QtAwesome==0.7.2
qtconsole==4.7.5
QtPy==1.9.0
regex==2020.6.8
requests==2.28.1
rope==0.17.0
SecretStorage==3.1.2
Send2Trash==1.5.0
setuptools==49.2.0
simple-multiprocessing==0.0.15
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.15.0
sniffio==1.3.0
snowballstemmer==2.0.0
sortedcollections==1.2.1
sortedcontainers==2.2.2
soupsieve==2.0.1
Sphinx==3.1.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sphinxcontrib-websupport==1.2.3
spyder==4.1.4
spyder-kernels==1.9.2
SQLAlchemy==1.3.18
starlette==0.13.6
toml==0.10.1
toolz==0.10.0
tornado==6.0.4
tqdm==4.47.0
traitlets==4.3.3
typing-extensions==4.4.0
ujson==1.35
unicodecsv==0.14.1
unicorn==2.0.1
urllib3==1.25.9
uvicorn==0.12.3
watchdog==0.10.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
wheel==0.34.2
widgetsnbextension==3.5.1
wrapt==1.11.2
wurlitzer==2.0.1
xlrd==1.2.0
XlsxWriter==1.2.9
xlwt==1.3.0
xmltodict==0.12.0
yapf==0.30.0
zict==2.0.0
zipp==3.1.0
